pyproject.toml
[project]
name = "langmem-examples"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "langgraph-api>=0.4.20",
    "langgraph-cli>=0.4.2",
    "langgraph-sdk>=0.2.8",
    "langmem>=0.0.29",
    "python-dotenv>=1.1.1",
    "yfinance>=0.2.66",
]
